+++ static/js/settings_config.ts
@@ -87,6 +87,7 @@ export const get_all_display_settings = (): DisplaySettings => ({
         user_display_settings: [
             "dense_mode",
             "high_contrast_mode",
+            "display_organization_name",
             "left_side_userlist",
             "fluid_layout_width",
             "starred_message_counts",
@@ -556,6 +557,7 @@ export const display_settings_labels = {
     dense_mode: $t({defaultMessage: "Dense mode"}),
     fluid_layout_width: $t({defaultMessage: "Use full width on wide screens"}),
     high_contrast_mode: $t({defaultMessage: "High contrast mode"}),
+    display_organization_name: $t({defaultMessage: "Display organization name in left sidebar"}),
     left_side_userlist: $t({
         defaultMessage: "Show user list on left sidebar in narrow windows",
     }),